Revert "Plan 9 from Bell Labs Second Edition 1995-04-05"
This reverts commit 219b2ee8daee37f4aad58d63f21287faa8e4ffdc.

+++ sys/src/cmd/map/libmap/azequalarea.c
@@ -1,7 +1,7 @@
 #include "map.h"
 
 int
-Xazequalarea(struct place *place, double *x, double *y)
+Xazequalarea(struct place *place, float *x, float *y)
 {
 	double r;
 	r = sqrt(1. - place->nlat.s);
